requirements.txt
Flask
Flask-Login
Flask-Mail
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn